pyproject.toml
[project]
name = "georag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "geopandas>=1.1.3",
    "numpy>=2.4.3",
    "ollama>=0.6.1",
    "pgvector>=0.4.2",
    "psycopg>=3.3.3",
    "psycopg-binary>=3.3.3",
    "pydantic-ai>=1.73.0",
    "pystac>=1.14.3",
    "pystac-client>=0.9.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
